Foreign

+++ app/Http/Controllers/AdminControllers/UserController.php
@@ -155,11 +155,9 @@ class UserController extends Controller
             DB::table('todos')->where('assignedTo', $id)->delete();
         };
         if ($user != null) {
-            /*if ($user->image != null) {
+            if ($user->image != null) {
                 unlink(public_path() . '/Uploads/users/thumbnails/' . $user->image);
-            }*/
-            /*$destroy= $todo->delete();
-            dd($destroy);*/
+            }
             $destroy = $user->delete();
             if ($destroy) {
                 return redirect()->route('user.index')->with('success', 'User deleted successfully');
